link plugin init script when add or update

+++ qwer-ps1.zsh
@@ -13,7 +13,7 @@ _qwer_ps1_usage() {
   # echo "         qwer-ps1           plugin update (--all)" >&2
   echo "         qwer-ps1           plugin remove <name>       -- remove plugin" >&2
   echo "         qwer-ps1           plugin is-installed <name> -- check whether the plugin is installed" >&2
-  echo "         qwer-ps1           init                       -- function to initialize qwer-ps1" >&2
+  echo "         qwer-ps1           init                       -- function to initialize qwer-ps1 and plugins" >&2
   echo "" >&2
   echo "options: -b <brackets pair (default: '[]')> -- brackets you want to use" >&2
   echo "         -c <color (default: 'red')>        -- color you want to use" >&2
@@ -65,7 +65,27 @@ _qwer_ps1_plugin_add() {
   local url="$2"
 
   git clone $url ${QWER_PS1_PLUGINS}/${name}
-  ln -s ${QWER_PS1_PLUGINS}/${name}/src/show-current ${QWER_PS1_SHIMS}/show-current-${name}
+  echo
+  if [[ ! -f ${QWER_PS1_PLUGINS}/${name}/src/show-current ]]; then
+    echo "${QWER_PS1_PLUGINS}/${name}/src/show-current not found. Invalid qwer-ps1 plugin..." >&2
+    return 1
+  fi
+
+  _qwer_ps1_plugin_link $name
+}
+
+_qwer_ps1_plugin_link() {
+  local name="$1"
+
+  if [[ ! -f ${QWER_PS1_SHIMS}/show-current-${name} ]]; then
+    echo "Linking ${QWER_PS1_PLUGINS}/${name}/src/show-current ..."
+    ln -s ${QWER_PS1_PLUGINS}/${name}/src/show-current ${QWER_PS1_SHIMS}/show-current-${name}
+  fi
+
+  if [[ -f ${QWER_PS1_PLUGINS}/${name}/src/init && ! -f ${QWER_PS1_SHIMS}/init-${name} ]]; then
+    echo "Linking ${QWER_PS1_PLUGINS}/${name}/src/init ..."
+    ln -s ${QWER_PS1_PLUGINS}/${name}/src/init ${QWER_PS1_SHIMS}/init-${name}
+  fi
 }
 
 _qwer_ps1_plugin_list() {
@@ -78,6 +98,8 @@ _qwer_ps1_plugin_update() {
   if [[ -d ${QWER_PS1_PLUGINS}/${name} ]]; then
     pushd ${QWER_PS1_PLUGINS}/${name}
     git pull
+    echo
+    _qwer_ps1_plugin_link $name
     popd
   else
     echo "Plugin $name not installed." >&2
